move controller components into folder
